Create code
In this example, the function nextSolarEclipse uses the Date object in JavaScript to get the current date, and then uses a while loop to increment the year until it finds the next solar eclipse date that is in the future. The date of the eclipse is calculated using the calculateSolarEclipse(year) function, which you would need to implement to suit your needs.

It's important to note that this is just an example, and it will not give you the exact date of the next solar eclipse, you will need to use a library or an API like NASA's API for that purpose.